+++ R/support.R
@@ -27,7 +27,6 @@ aggregate <- function(x, what, as_incid = FALSE, stratum = NULL,
 ##' @return cumulative and incident flows
 ##' @export
 extract_flows_xpvwrh <- function(y) {
-
 # extract cumulative flows
   flow_names <- c("cum_diag_a", "cum_diag_s", "cum_treated", "cum_screened",
                   "cum_vaccinated", "cum_vbe")
@@ -35,7 +34,6 @@ extract_flows_xpvwrh <- function(y) {
   names(cumulative_flows) <- flow_names
 
 ## extract vaccinations and revaccinations separately
-
   # primary vaccinated = everyone vaccinated from X(1) regardless of # doses
   cumulative_flows$cum_primary_total <-
     t(aggregate(y, "cum_vaccinated", stratum = 1))