+++ tests/testthat/test-run_grid.R
@@ -153,6 +153,110 @@ test_that("compare baseline xpvwrh works as expected", {
   expect_equal(calc_inc_costs(85, cost) - calc_inc_costs(70, cost),
                cost$pv_inc_doses * (85 - 70))
 
+  ## test that proportion of the population vaccine protected is calculated
+  # correctly
+
+   # when uptake = 0, vaccine protected raw and proportion = 0
+   y <- run_onevax_xpvwrh(tt, gp,
+                         r1 = 0,
+                         r2 = 0, booster_uptake = 0,
+                         strategy = "VoD")
+   bl <- extract_flows_xpvwrh(run_onevax_xpvwrh(tt, gp, ip))
+
+   z <- compare_baseline_xpvwrh(y, bl, uptake_first_dose = 0,
+                               uptake_second_dose = 0, cp, 0)
+
+    expect_true(all(z$inc_vacprotec_total == 0))
+    expect_true(all(z$inc_vacprotec_total_prop == 0))
+
+   # when r2 = 1, vac_protect_part = 0
+
+   y <- run_onevax_xpvwrh(tt, gp,
+                           r1 = 0.5,
+                           r2 = 1, booster_uptake = 0,
+                           strategy = "VoD")
+
+   z <- compare_baseline_xpvwrh(y, bl, uptake_first_dose = 0.5,
+                                 uptake_second_dose = 0.2, cp, 0)
+
+  expect_true(all(z$vacprotec_part == 0))
+  expect_true(all(z$vacprotec_part_prop == 0))
+  expect_true(all(z$vacprotec_total != 0))
+  expect_true(all(z$vacprotec_total_prop != 0))
+  expect_equal(z$vacprotec_total, z$vacprotec_full)
+
+  # proportion * 6e05 = raw number
+  expect_equal(z$vacprotec_total_prop * 6e05, z$vacprotec_total)
+
+  N <- t(aggregate(y, "N"))[1, 1]
+  expect_equal(z$vacprotec_total_prop * N, z$vacprotec_total)
+
+  # when model is initiated with a certain % vaccine coverage, this
+  # is the model proportion calculated
+
+  # move % pop into P for starting condition, no uptake, no waning
+  # expect the correct proportion under partial vac
+  tt <- seq(0, 5)
+  gp <- gono_params(1:2)
+  pars <- lapply(gp[1], model_params)
+  ip <- lapply(pars, initial_params_xpvwrh, coverage_v = 0.1, coverage_p = 0.5,
+               t = 5)
+
+  # get number of people in P & V
+  n_p <- sum(ip[[1]]$U0[, 2])
+
+  n_v <- sum(ip[[1]]$U0[, 3])
+
+  y_init <- run_onevax_xpvwrh(tt, gp, init_params = ip, dur_v = 1e+90)
+
+  bl_2 <- extract_flows_xpvwrh(run_onevax_xpvwrh(tt, gp, ip))
+
+  # N.B as people die at a constant rate from all compartments across all strata
+  # for the output of compare_baselines_xpvwrh, z,
+  # z$vacprotec_part/full/total for the first timepoint will not give the value
+  # corresponding to the number of people initially vaccinated. E.g: We expect
+  # for strata P that initial coverage of partial vaccination was 50% but some
+  # will have died between t = 0 an t = 1 so compare_baselines won't output 3e05
+  # , as part of its code removes the timepoint 0 from the output.
+  # - instead we will copy the code from compare_baselines_xpvwrh that generates
+  # vac_protec_part/full/total but not ommit t = 0 as we usually do and see
+  # if this gives us the values we expect:
+
+             ## extract number under vaccine protection
+             vacsnap <- list()
+             # fully vaccine protected, snapshot of N in: V(3) and R(5)
+             vacsnap$vacprotec_full <-
+               t(aggregate(y_init, "N", stratum = c(3, 5)))
+
+             # partially vaccine protected, snapshot of N in: P(2)
+             vacsnap$vacprotec_part <-
+               t(aggregate(y_init, "N", stratum = 2))
+
+             # vaccine protected total, snapshot of N in: P(2), V(3), W(4)
+             vacsnap$vacprotec_total <-
+               t(aggregate(y_init, "N", stratum = c(2, 3, 5)))
+
+             ## calculate proportion of the population under vaccine protection
+             # get total pop size  (including H!) This should be the same across
+             # all model runs for all timepoints
+             N <- t(aggregate(y, "N"))[1, 1]
+
+              vacsnap$vacprotec_full_prop <- vacsnap$vacprotec_full / N
+              vacsnap$vacprotec_part_prop <- vacsnap$vacprotec_part / N
+              vacsnap$vacprotec_total_prop <- vacsnap$vacprotec_total / N
+
+             # calculate proportion population vaccinated initially
+             # from the initial conditions object used for the model run
+             init_vac_p <- n_p / N
+             init_vac_v <- n_v / N
+
+        for (i in seq_along(y_init)) {
+        expect_equal(vacsnap$vacprotec_part_prop[1, i], init_vac_p)
+        expect_equal(vacsnap$vacprotec_full_prop[1, i], init_vac_v)
+        expect_equal(vacsnap$vacprotec_total_prop[1, i],
+                     init_vac_p + init_vac_v)
+        }
+
   ## test that cumulative primary and booster vaccination calc correctly
 
   bl <- extract_flows_xpvwrh(run_onevax_xpvwrh(tt, gp))